Create ExcelCheckBoxAlternative
How to use:

Open the example workbook and activate the sheet "Cells as CheckBoxes". Doubleclick on any of the indicated cells. The cells will toggle between a "check" and null. Now activate the sheet "Mutually Exclusive examples" and double click on different indicated cells. Note the effect of doubleclicks in each color group.
Test the code:

Right Click on the sheet tab and choose "View code". Paste the code for the worksheet "Cells as Checkboxes" in the panel. Select the cells you want to specify as "checkbox" cells and give them a name ("myChecks" in example). You can select non-contiguous cells by holding down the CTRL key. Now when you double click on one of these cells, it will toggle between a "check" and null.
For the second part, paste the code for the worksheet "Mutually Exclusive example" in the worksheet module as described above. Again, select the cells to specify and name them. The code on this sheet is set for a named range "Ckboxes". Also, the Change Events are cell specific. You can modify the ranges if needed.
Hint: Deselect "Edit directly in cells" using Tools>Options>Edit Tab